requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
elasticsearch==8.15.0
langchain-ollama==0.2.0
langchain-openai==0.2.14
langchain-core==0.3.27
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
mistune==3.0.2
aiohttp==3.11.11
cohere==5.11.4
openai==1.58.1
tiktoken==0.8.0
sentence-transformers>=2.2.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# Email
aiosmtplib>=3.0.0
email-validator>=2.1.0



